ヘッダーコンポーネントのファイル名を変更
