+++ client/layouts/blog.vue
@@ -1,6 +1,10 @@
 <template>
   <div class="wrap">
-    <Header />
+    <header class="l-header">
+      <h1>
+        <logo />
+      </h1>
+    </header>
     <nuxt />
     <div class="footer_wrap">
       <nui-footer />
@@ -15,18 +19,40 @@
   min-height: 100vh;
 }
 
+.l-header {
+  background-color: $color-white1;
+  padding: 10px 0 2px;
+  text-align: center;
+}
+
+@media screen and (min-width: 769px) {
+  .l-header {
+    padding: 17px 0 10px;
+  }
+}
+
+.l-header-logo {
+  width: 170px;
+}
+
+@media screen and (min-width: 769px) {
+  .l-header-logo {
+    width: 222px;
+  }
+}
+
 .footer_wrap {
   margin-top: auto;
 }
 </style>
 
 <script>
-import Header from "@/components/layouts/Header.vue"
+import Logo from "@/components/svg/Logo.vue"
 import NuiFooter from "@/components/layouts/Footer.vue"
 
 export default {
   components: {
-    Header,
+    Logo,
     NuiFooter,
   },
 }
